Rename config/bulk.py to config.py
